src/shared/components/palette/linked-cases/components/check-your-answers/check-your-answers.component.ts
import { Component, EventEmitter, Output } from '@angular/core';
import { ErrorMessage } from '../../../../../domain';
import { LinkedCasesState } from '../../domain';
import { LinkedCasesPages } from '../../enums';

@Component({
  selector: 'ccd-linked-cases-check-your-answers',
  templateUrl: './check-your-answers.component.html'
})
export class CheckYourAnswersComponent  {

  @Output()
  public linkedCasesStateEmitter: EventEmitter<LinkedCasesState> = new EventEmitter<LinkedCasesState>();

  errorMessages: ErrorMessage[];

  public onNext(): void {
    // Return linked cases state and error messages to the parent
    this.linkedCasesStateEmitter.emit({ currentLinkedCasesPage: LinkedCasesPages.CHECK_YOUR_ANSWERS, errorMessages: this.errorMessages });
  }
}
